+++ src/pages/protected/article/CreateArticle/CreateArticle.jsx
@@ -1,122 +1,119 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import * as yup from 'yup';
-import { useForm, Controller } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
+import React, { useState, useRef, useEffect } from 'react';
 import {
   Container,
-  TextField,
-  Button,
-  Select,
-  MenuItem,
-  Box,
-  FormHelperText,
   Stack,
   Typography,
-  CircularProgress,
-  FormControl,
+  TextField,
+  Button,
+  FormHelperText,
+  Select,
+  MenuItem,
   InputLabel,
+  FormControl,
+  CircularProgress,
 } from '@mui/material';
-import { omit } from 'ramda';
-import MDEditor from '@uiw/react-md-editor';
 import { useDispatch, useSelector } from 'react-redux';
-import { getCategories } from '../../../store/features/category/categoryMiddlewares';
-import { formMargin, ROUTES } from '../../../helpers';
-import { uploadImage } from '../../../store/features/image/imageMiddlewares';
-import { getArticleByUrl, updateArticle } from '../../../store/features/article/articleMiddlewares';
-
-const ArticleDto = (article) => {
-  return {
-    title: article?.title || '',
-    url: article?.url || '',
-    spoiler: article?.spoiler || '',
-    content: article?.content || '',
-    category: article?.category._id || '',
-    coverImage: article?.coverImage || '',
-  };
-};
+import { useForm, Controller } from 'react-hook-form';
+import { yupResolver } from '@hookform/resolvers/yup';
+import { omit } from 'ramda';
+import * as yup from 'yup';
+import MDEditor from '@uiw/react-md-editor';
+import { createArticle } from '../../../../store/features/article/articleMiddlewares';
+import { getCategories } from '../../../../store/features/category/categoryMiddlewares';
+import { uploadImage } from '../../../../store/features/image/imageMiddlewares';
+import { DEFAULT_ARTICLE_IMAGE } from '../../../../helpers/constants/constans';
+import { ERROR_MESSAGES, HELPER_TEXT } from '../../../../helpers';
 
 const validationSchema = yup
   .object({
-    title: yup.string().max(255).required(),
-    url: yup.string().max(255).required(),
-    spoiler: yup.string().max(1000).required(),
-    content: yup.string().max(10000).required(),
+    title: yup.string().required().max(255),
+    category: yup.string().required(),
+    url: yup.string().required().max(30),
+    spoiler: yup.string().required().max(100),
+    content: yup.string().required().max(10000),
+    picture: yup.string(),
   })
   .required();
 
-export const UpdateArticle = () => {
-  const { newsUrl } = useParams();
-  const navigate = useNavigate();
+export const CreateArticle = () => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [categories, setCategories] = useState([]);
   const [serverError, setServerError] = useState('');
+  const { userInfo } = useSelector((state) => state.auth);
   const formRef = useRef();
-
   const dispatch = useDispatch();
-  const { categories } = useSelector((state) => state.category);
-  const { article, loadingArticle } = useSelector((state) => state.article);
 
   const {
     handleSubmit,
-    control,
-    reset,
     formState: { errors },
+    reset,
+    control,
   } = useForm({
-    defaultValues: ArticleDto(),
+    defaultValues: {
+      title: '',
+      category: '',
+      url: '',
+      spoiler: '',
+      content: '',
+    },
     resolver: yupResolver(validationSchema),
   });
 
   useEffect(() => {
-    dispatch(getArticleByUrl({ newsUrl }));
-    dispatch(getCategories({ isActive: true }));
-    reset(ArticleDto(article));
-  }, []);
+    if (categories.length === 0) {
+      dispatch(getCategories({ isActive: true })).then((res) => {
+        setCategories(res.payload);
+      });
+    }
+  }, [categories.length, dispatch]);
 
   const onSubmit = async (data) => {
     try {
-      setServerError('');
       const formData = new FormData(formRef.current);
-      const image = formData.get('image');
+      const image = formData.get('coverImage');
 
       setServerError('');
+      setIsLoading(true);
       await dispatch(uploadImage({ image })).then((res) => {
         dispatch(
-          updateArticle({
-            id: article._id,
+          createArticle({
             ...data,
-            coverImage: res.payload.data || article.coverImage,
+            author: userInfo._id,
+            coverImage: res.payload.data || DEFAULT_ARTICLE_IMAGE,
           }),
         );
       });
-      navigate(ROUTES.HOME);
+      reset();
     } catch (err) {
-      setServerError('Server Error');
+      setServerError(ERROR_MESSAGES.SERVER_ERROR);
+      console.log(err);
+    } finally {
+      setIsLoading(false);
     }
   };
 
-  const isButtonDisabled = Object.keys(errors).length > 0;
+  const isButtonDisabled = Object.keys(errors).length > 0 || isLoading;
 
   return (
-    article &&
-    categories && (
+    <>
       <Container maxWidth="md">
         <Stack ref={formRef} component="form" spacing={2} onSubmit={handleSubmit(onSubmit)}>
           <Typography variant="h4" variantMapping={{ h4: 'h1' }} gutterBottom>
-            Oновіть дані статті
+            Введіть дані статті
           </Typography>
+          {!!serverError && <FormHelperText error>{serverError}</FormHelperText>}
+
           <Controller
             name="title"
             control={control}
             render={({ field }) => (
-              <Box sx={formMargin}>
-                <TextField
-                  margin="normal"
-                  label="Заголовок:"
-                  fullWidth
-                  {...field}
-                  error={!!errors?.title}
-                  helperText={!!errors.title?.message}
-                />
-              </Box>
+              <TextField
+                fullWidth
+                label="Заголовок"
+                error={!!errors.title}
+                helperText={errors.title?.message ? errors.title.message : HELPER_TEXT.TITLE_TIP}
+                {...field}
+              />
             )}
           />
 
@@ -153,36 +150,34 @@ export const UpdateArticle = () => {
             name="url"
             control={control}
             render={({ field }) => (
-              <Box sx={formMargin}>
-                <TextField
-                  margin="normal"
-                  label="Url:"
-                  fullWidth
-                  {...field}
-                  error={!!errors?.url}
-                  helperText={!!errors.url?.message}
-                />
-              </Box>
+              <TextField
+                fullWidth
+                label="Url:"
+                error={!!errors.url}
+                helperText={errors.url?.message ? errors.url.message : HELPER_TEXT.URL_TIP}
+                {...field}
+              />
             )}
           />
+
           <Controller
             name="spoiler"
             control={control}
             render={({ field }) => (
-              <Box sx={formMargin}>
-                <TextField
-                  multiline
-                  label="Спойлер:"
-                  rows={3}
-                  margin="normal"
-                  fullWidth
-                  {...field}
-                  error={!!errors?.spoiler}
-                  helperText={!!errors.spoiler?.message}
-                />
-              </Box>
+              <TextField
+                fullWidth
+                multiline
+                rows={3}
+                label="Спойлер"
+                error={!!errors.spoiler}
+                helperText={
+                  errors.spoiler?.message ? errors.spoiler.message : HELPER_TEXT.SPOILER_TIP
+                }
+                {...field}
+              />
             )}
           />
+
           <Controller
             name="content"
             control={control}
@@ -204,7 +199,7 @@ export const UpdateArticle = () => {
                 <>
                   <input hidden type="file" accept="image/*" id="coverImage" {...fieldProps} />
                   <Button htmlFor="coverImage" variant="contained" component="label" fullWidth>
-                    Завантажити обкладинкy
+                    Завантажити обкладинку
                   </Button>
                   {errors.coverImage && (
                     <FormHelperText error>{errors.coverImage?.message}</FormHelperText>
@@ -214,17 +209,16 @@ export const UpdateArticle = () => {
             }}
           />
 
-          {!!serverError && <FormHelperText error>{serverError}</FormHelperText>}
           <Button
             type="submit"
             variant="contained"
             disabled={isButtonDisabled}
-            startIcon={loadingArticle && <CircularProgress size={20} />}
+            startIcon={isLoading && <CircularProgress size={20} sx={{ color: 'white' }} />}
           >
-            Зберегти статтю
+            Зберегти
           </Button>
         </Stack>
       </Container>
-    )
+    </>
   );
 };
