+++ src/pages/protected/category/UpdateCategories/UpdateCategories.jsx
@@ -2,11 +2,11 @@ import { List, Container, Button } from '@mui/material';
 import React from 'react';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { ItemCategory } from '../../../components/category/ItemCategory/ItemCategory';
-import { getCategories } from '../../../store/features/category/categoryMiddlewares';
+import { ItemCategory } from '../../../../components/category/ItemCategory/ItemCategory';
+import { getCategories } from '../../../../store/features/category/categoryMiddlewares';
 import AddIcon from '@mui/icons-material/Add';
 import { useNavigate } from 'react-router-dom';
-import { ROUTES } from '../../../helpers';
+import { ROUTES } from '../../../../helpers';
 
 export const UpdateCategories = () => {
   const { categories } = useSelector((state) => state.category);
