+++ src/pages/protected/comment/UnpublishedComments/UnpublishedComments.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container } from '@mui/material';
-import { CommentsList } from '../../../components/comment/CommentsList/CommentsList';
-import { PAGE_TYPE } from '../../../helpers';
+import { CommentsList } from '../../../../components/comment/CommentsList/CommentsList';
+import { PAGE_TYPE } from '../../../../helpers';
 
 export const UnpublishedComments = () => {
   return (
