+++ src/pages/protected/user/UserPage/UserPage.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Container } from '../../../components/layout/Container/Container';
+import { Container } from '../../../../components/layout/Container/Container';
 import { useDispatch, useSelector } from 'react-redux';
-import { UserCard } from '../../../components/user/UserCard/UserCard';
+import { UserCard } from '../../../../components/user/UserCard/UserCard';
 import { Box, CircularProgress } from '@mui/material';
 import { useEffect } from 'react';
-import { getUser } from '../../../store/features/auth/authMiddlewares';
+import { getUser } from '../../../../store/features/auth/authMiddlewares';
 
 export const UserPage = () => {
   const { userInfo, accessToken, loading } = useSelector((state) => state.auth);
